requirements/prod.txt
Django>=4.2,<4.3
django-dotenv>=1.4,<1.5
django-mptt==0.15.0
djangorestframework>=3.14,<3.15
drf-spectacular==0.26.5
psycopg2>=2.9,<3
